docs: 불필요한 패키지 제거 및 테스트 클래스 패키지 분리
