+++ README.md
@@ -38,17 +38,3 @@
     - [x] 뒤로는 갈 수 없다.
   - [x] 포 : 앞에 기물이 한 개 존재할 때 원하는 곳으로 뛰어넘을 수 있다.
     - [x] 포끼리는 뛰어넘을 수 없다.
-
-## 출력
-- [ ] 초기화된 장기판을 출력한다.
-
-```
-          0  1  2  3   4  5  6  7  8
-      0  [차][마][상][사] . [사][상][마][차]
-      1   .  .  .  . [장] .  .  .  .
-      2   . [포] .  .  .  .  . [포] .
-      3  [졸] . [졸] . [졸] . [졸] .  [졸]
-      ...
-      9  초나라 기물들
-```
-
