--- src/test/java/janggiBoard/PawnTest.java
@@ -1,4 +1,0 @@
-package janggiBoard;
-
-public class PawnTest {
-}
