+++ src/test/java/janggiBoard/PieceTest/ElephantTest.java
@@ -1,48 +1,46 @@
-package janggiBoard;
+package janggiBoard.PieceTest;
 
 import domain.Position;
 import domain.Team;
-import domain.piece.Blank;
-import domain.piece.Horse;
-import domain.piece.Piece;
-import domain.piece.PieceProvider;
+import domain.piece.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class HorseTest {
+public class ElephantTest {
+
     private TestPieceProvider testBoard;
-    private Horse horse;
+    private Elephant elephant;
 
     @BeforeEach
     void setUp() {
-        horse = new Horse(Team.CHO);
+        elephant = new Elephant(Team.CHO);
         testBoard = new TestPieceProvider();
     }
 
     @Test
     void 마가_목적지에_갈_수_있다() {
         Position currentPosition = new Position(5, 5);
-        Position targetPosition = new Position(3, 4);
+        Position targetPosition = new Position(3, 8);
 
         testBoard.setAllBlank();
-        boolean isCanMove = horse.canMove(currentPosition, targetPosition, testBoard);
+        boolean isCanMove = elephant.canMove(currentPosition, targetPosition, testBoard);
         assertThat(isCanMove).isTrue();
     }
 
     @Test
     void 마가_목적지에_갈_수_없다() {
         Position currentPosition = new Position(5, 5);
-        Position targetPosition = new Position(3, 4);
+        Position targetPosition = new Position(3, 8);
 
         testBoard.setAllBlank();
-        testBoard.setBlank(new Position(4, 5));
+        testBoard.setBlank(new Position(5, 6));
 
-        boolean isCanMove = horse.canMove(currentPosition, targetPosition, testBoard);
+        boolean isCanMove = elephant.canMove(currentPosition, targetPosition, testBoard);
         assertThat(isCanMove).isFalse();
     }
