+++ src/test/java/janggiBoard/PieceTest/HorseTest.java
@@ -1,44 +1,49 @@
-package janggiBoard;
+package janggiBoard.PieceTest;
 
 import domain.Position;
+import domain.Team;
 import domain.piece.Blank;
+import domain.piece.Horse;
 import domain.piece.Piece;
 import domain.piece.PieceProvider;
-import domain.strategy.PalaceStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.*;
 
-public class PalaceStrategyTest {
-
-    private PalaceStrategy palaceStrategy;
+public class HorseTest {
     private TestPieceProvider testBoard;
+    private Horse horse;
 
     @BeforeEach
     void setUp() {
-        palaceStrategy = new PalaceStrategy();
+        horse = new Horse(Team.CHO);
         testBoard = new TestPieceProvider();
     }
 
     @Test
-    void 궁과_사가_주변에_장애물이_없다면_8가지_후보_모두_반환() {
+    void 마가_목적지에_갈_수_있다() {
         Position currentPosition = new Position(5, 5);
+        Position targetPosition = new Position(3, 4);
+
         testBoard.setAllBlank();
+        boolean isCanMove = horse.canMove(currentPosition, targetPosition, testBoard);
+        assertThat(isCanMove).isTrue();
+    }
 
-        List<Position> candidates = palaceStrategy.getMoveCandidates(currentPosition, testBoard);
+    @Test
+    void 마가_목적지에_갈_수_없다() {
+        Position currentPosition = new Position(5, 5);
+        Position targetPosition = new Position(3, 4);
 
-        assertThat(candidates).hasSize(8)
-                .containsExactlyInAnyOrder(
-                        new Position(4, 5), new Position(4, 4),
-                        new Position(4, 6), new Position(5, 4),
-                        new Position(5, 6), new Position(6, 4),
-                        new Position(6, 5), new Position(6, 6)
-                );
+        testBoard.setAllBlank();
+        testBoard.setBlank(new Position(4, 5));
+
+        boolean isCanMove = horse.canMove(currentPosition, targetPosition, testBoard);
+        assertThat(isCanMove).isFalse();
     }
 
     private static class TestPieceProvider implements PieceProvider {
