+++ src/test/java/janggiBoard/PieceTest/PawnTest.java
@@ -1,0 +1,4 @@
+package janggiBoard.PieceTest;
+
+public class PawnTest {
+}
