+++ src/test/java/janggiBoard/strategyTest/CarStrategyTest.java
@@ -1,4 +1,4 @@
-package janggiBoard;
+package janggiBoard.strategyTest;
 
 import domain.Position;
 import domain.piece.Blank;
