+++ src/test/java/janggiBoard/strategyTest/ElephantStrategyTest.java
@@ -1,11 +1,10 @@
-package janggiBoard;
-
+package janggiBoard.strategyTest;
 
 import domain.Position;
 import domain.piece.Blank;
 import domain.piece.Piece;
 import domain.piece.PieceProvider;
-import domain.strategy.HorseStrategy;
+import domain.strategy.ElephantStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -13,31 +12,32 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class HorseStrategyTest {
-    private HorseStrategy horseStrategy;
+public class ElephantStrategyTest {
+
+    private ElephantStrategy elephantStrategy;
     private TestPieceProvider testBoard;
 
     @BeforeEach
-    void setUp() {
-        horseStrategy = new HorseStrategy();
+    public void setUp() {
+        elephantStrategy = new ElephantStrategy();
         testBoard = new TestPieceProvider();
     }
 
     @Test
-    void 마_주변에_장애물_없으면_8가지_후보_모두_반환() {
+    void 상_주변에_장애물_없으면_8가지_후보_모두_반환() {
         Position position = new Position(5, 5);
         testBoard.setAllBlank();
 
-        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(8)
                 .containsExactlyInAnyOrder(
-                        new Position(3, 4), new Position(3, 6), // 북쪽 기반
-                        new Position(7, 4), new Position(7, 6), // 남쪽 기반
-                        new Position(4, 3), new Position(6, 3), // 서쪽 기반
-                        new Position(4, 7), new Position(6, 7)  // 동쪽 기반
+                        new Position(2, 3), new Position(2, 7), // 북쪽 기반
+                        new Position(3, 8), new Position(7, 8), // 남쪽 기반
+                        new Position(8, 3), new Position(8, 7), // 서쪽 기반
+                        new Position(3, 2), new Position(7, 2)  // 동쪽 기반
                 );
     }
 
@@ -48,10 +48,10 @@ public class HorseStrategyTest {
 
         // 북쪽 멱 위치를 막힌 상태로 설정
         testBoard.setBlank(new Position(4, 5));
-        List<Position> candidates = horseStrategy.getMoveCandidates(source, testBoard);
+        List<Position> candidates = elephantStrategy.getMoveCandidates(source, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(3, 4), new Position(3, 6));
+                .doesNotContain(new Position(2, 3), new Position(2, 7));
     }
 
     @Test
@@ -60,10 +60,10 @@ public class HorseStrategyTest {
         testBoard.setAllBlank();
 
         testBoard.setBlank(new Position(6, 5));
-        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(7,4), new Position(7,6));
+                .doesNotContain(new Position(8,3), new Position(8,7));
     }
 
     @Test
@@ -72,10 +72,10 @@ public class HorseStrategyTest {
         testBoard.setAllBlank();
 
         testBoard.setBlank(new Position(5, 4));
-        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(6,5), new Position(4,3));
+                .doesNotContain(new Position(3,2), new Position(7,2));
     }
 
     @Test
@@ -84,10 +84,10 @@ public class HorseStrategyTest {
         testBoard.setAllBlank();
 
         testBoard.setBlank(new Position(5, 6));
-        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(4, 7), new Position(6, 7));
+                .doesNotContain(new Position(3, 8), new Position(7, 8));
     }
 
     private static class TestPieceProvider implements PieceProvider {
