+++ src/test/java/janggiBoard/strategyTest/HorseStrategyTest.java
@@ -1,10 +1,11 @@
-package janggiBoard;
+package janggiBoard.strategyTest;
+
 
 import domain.Position;
 import domain.piece.Blank;
 import domain.piece.Piece;
 import domain.piece.PieceProvider;
-import domain.strategy.ElephantStrategy;
+import domain.strategy.HorseStrategy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -12,32 +13,31 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
-public class ElephantStrategyTest {
-
-    private ElephantStrategy elephantStrategy;
+public class HorseStrategyTest {
+    private HorseStrategy horseStrategy;
     private TestPieceProvider testBoard;
 
     @BeforeEach
-    public void setUp() {
-        elephantStrategy = new ElephantStrategy();
+    void setUp() {
+        horseStrategy = new HorseStrategy();
         testBoard = new TestPieceProvider();
     }
 
     @Test
-    void 상_주변에_장애물_없으면_8가지_후보_모두_반환() {
+    void 마_주변에_장애물_없으면_8가지_후보_모두_반환() {
         Position position = new Position(5, 5);
         testBoard.setAllBlank();
 
-        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(8)
                 .containsExactlyInAnyOrder(
-                        new Position(2, 3), new Position(2, 7), // 북쪽 기반
-                        new Position(3, 8), new Position(7, 8), // 남쪽 기반
-                        new Position(8, 3), new Position(8, 7), // 서쪽 기반
-                        new Position(3, 2), new Position(7, 2)  // 동쪽 기반
+                        new Position(3, 4), new Position(3, 6), // 북쪽 기반
+                        new Position(7, 4), new Position(7, 6), // 남쪽 기반
+                        new Position(4, 3), new Position(6, 3), // 서쪽 기반
+                        new Position(4, 7), new Position(6, 7)  // 동쪽 기반
                 );
     }
 
@@ -48,10 +48,10 @@ public class ElephantStrategyTest {
 
         // 북쪽 멱 위치를 막힌 상태로 설정
         testBoard.setBlank(new Position(4, 5));
-        List<Position> candidates = elephantStrategy.getMoveCandidates(source, testBoard);
+        List<Position> candidates = horseStrategy.getMoveCandidates(source, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(2, 3), new Position(2, 7));
+                .doesNotContain(new Position(3, 4), new Position(3, 6));
     }
 
     @Test
@@ -60,10 +60,10 @@ public class ElephantStrategyTest {
         testBoard.setAllBlank();
 
         testBoard.setBlank(new Position(6, 5));
-        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(8,3), new Position(8,7));
+                .doesNotContain(new Position(7,4), new Position(7,6));
     }
 
     @Test
@@ -72,10 +72,10 @@ public class ElephantStrategyTest {
         testBoard.setAllBlank();
 
         testBoard.setBlank(new Position(5, 4));
-        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(3,2), new Position(7,2));
+                .doesNotContain(new Position(6,5), new Position(4,3));
     }
 
     @Test
@@ -84,10 +84,10 @@ public class ElephantStrategyTest {
         testBoard.setAllBlank();
 
         testBoard.setBlank(new Position(5, 6));
-        List<Position> candidates = elephantStrategy.getMoveCandidates(position, testBoard);
+        List<Position> candidates = horseStrategy.getMoveCandidates(position, testBoard);
 
         assertThat(candidates).hasSize(6)
-                .doesNotContain(new Position(3, 8), new Position(7, 8));
+                .doesNotContain(new Position(4, 7), new Position(6, 7));
     }
 
     private static class TestPieceProvider implements PieceProvider {
